+++ controllers/auth.controllers.js
@@ -1,7 +1,8 @@
 const Model = require("../models/index");
 const { validationResult } = require("express-validator");
 const bcrypt = require("bcrypt");
-const jwt = require("jsonwebtoken");
+const GenerateToken = require("../util/GenerateToken");
+const validateToken = require("../routes/middlewares/validate");
 
 const createUser = async (req, res) => {
   const { firstName, lastName, email, password } = req.body;
@@ -19,7 +20,6 @@ const createUser = async (req, res) => {
     });
   }
 
-  const GenerateToken = require("../util/GenerateToken");
   // if there are no errors, save user in database
   if (errors[0].errors.length === 0) {
     const user = await Model.User.create({
@@ -42,4 +42,23 @@ const createUser = async (req, res) => {
   }
 };
 
-module.exports = { createUser };
+const authMe = async (req, res) => {
+    const token = req.headers.authorization;
+
+  if (!token) {
+    return res.status(401).json({
+      auth: false,
+      message: "No token provided",
+    });
+  }
+
+  const decoded = validateToken(token);
+  const user = await Model.User.findByPk(decoded.user);
+  res.json({
+    firstName: user.firstName,
+    lastName: user.lastName,
+    emai: user.email,
+  });
+};
+
+module.exports = { createUser, authMe};
