+++ routes/auth.js
@@ -2,33 +2,14 @@ const router = require("express").Router();
 const { User } = require("../models/index");
 const { body, validationResult } = require("express-validator");
 const bcrypt = require("bcrypt");
-const { jwt } = require("jsonwebtoken");
-const validateToken = require("./middlewares/validate");
-
-
+const { createUser, authMe } = require("../controllers/auth.controllers");
+const userValidationSchema = require("../validations/userValidationSchema");
 
 // auth me
-router.get("/me", async (req, res) => {
-  const token = req.headers.authorization;
+router.get("/me", authMe);
 
-  if (!token) {
-    return res.status(401).json({
-      auth: false,
-      message: "No token provided",
-    });
-  }
-
-  const decoded = validateToken(token);
-  const user = await User.findByPk(decoded.user);
-  res.json({
-    firstName: user.firstName,
-    lastName: user.lastName,
-    emai: user.email,
-  });
-});
-
-const { createUser } = require('../controllers/users.controllers');
-const userValidationSchema = require('../validations/userValidationSchema')
+// Create a new user
+router.post("/register", userValidationSchema, createUser);
 
 // login user
 router.post(
@@ -36,23 +17,19 @@ router.post(
   body("email").isEmail(),
   body("password").isLength({ min: 6 }),
   (req, res) => {
-    if (!validationResult(req).isEmpty())
+      if (!validationResult(req).isEmpty())
       return res.status(401).json({ ok: false });
-
-    const { email, password } = req.body;
-    User.findOne({
-      where: { email },
+      
+      const { email, password } = req.body;
+      User.findOne({
+          where: { email },
     }).then((user) => {
-      if (!user) return res.status(401).json({ ok: false });
-      if (!bcrypt.compareSync(password, user.password))
+        if (!user) return res.status(401).json({ ok: false });
+        if (!bcrypt.compareSync(password, user.password))
         return res.status(401).json({ ok: false });
-      return res.json({ ok: true, user });
+        return res.json({ ok: true, user });
     });
-  }
+}
 );
 
-
-
-// Create a new user
-router.post('/register', userValidationSchema, createUser)
 module.exports = router;
